testdriver/vbox*.py: Oh, don't be 'smart' and use pollTask to check whether the VM needs powering off.  Added needsPoweringOff to the session wrapper for doing that.
svn:sync-xref-src-repo-rev: r108394

+++ src/VBox/ValidationKit/testdriver/vbox.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# $Id: vbox.py 61950 2016-06-30 09:17:51Z knut.osmundsen@oracle.com $
+# $Id: vbox.py 61951 2016-06-30 09:33:48Z knut.osmundsen@oracle.com $
 # pylint: disable=C0302
 
 """
@@ -27,7 +27,7 @@ CDDL are applicable instead of those of the GPL.
 You may elect to license modified versions of this file under the
 terms and conditions of either the GPL or the CDDL or both.
 """
-__version__ = "$Revision: 61950 $"
+__version__ = "$Revision: 61951 $"
 
 
 # Standard Python imports.
@@ -2648,34 +2648,35 @@ class TestDriver(base.TestDriver):                                              
 
         # Check if the VM has terminated by it self before powering it off.
         fClose = True;
-        fRc = oSession.pollTask();
-        if fRc is not True:
+        fRc    = True;
+        if oSession.needsPoweringOff():
             reporter.log('terminateVmBySession: powering off "%s"...' % (oSession.sName,));
             fRc = oSession.powerOff(fFudgeOnFailure = False);
-        if fRc is not True:
-            # power off failed, try terminate it in a nice manner.
-            fRc = False;
-            uPid = oSession.getPid();
-            if uPid is not None:
-                reporter.error('terminateVmBySession: Terminating PID %u (VM %s)' % (uPid, oSession.sName));
-                fClose = base.processTerminate(uPid);
-                if fClose is True:
-                    self.waitOnDirectSessionClose(oSession.oVM, 5000);
-                    fClose = oSession.waitForTask(1000);
-
-                if fClose is not True:
-                    # Being nice failed...
-                    reporter.error('terminateVmBySession: Termination failed, trying to kill PID %u (VM %s) instead' \
-                                   % (uPid, oSession.sName));
-                    fClose = base.processKill(uPid);
+            if fRc is not True:
+                # power off failed, try terminate it in a nice manner.
+                fRc = False;
+                uPid = oSession.getPid();
+                if uPid is not None:
+                    reporter.error('terminateVmBySession: Terminating PID %u (VM %s)' % (uPid, oSession.sName));
+                    fClose = base.processTerminate(uPid);
                     if fClose is True:
                         self.waitOnDirectSessionClose(oSession.oVM, 5000);
                         fClose = oSession.waitForTask(1000);
+
                     if fClose is not True:
-                        reporter.error('terminateVmBySession: Failed to kill PID %u (VM %s)' % (uPid, oSession.sName));
+                        # Being nice failed...
+                        reporter.error('terminateVmBySession: Termination failed, trying to kill PID %u (VM %s) instead' \
+                                       % (uPid, oSession.sName));
+                        fClose = base.processKill(uPid);
+                        if fClose is True:
+                            self.waitOnDirectSessionClose(oSession.oVM, 5000);
+                            fClose = oSession.waitForTask(1000);
+                        if fClose is not True:
+                            reporter.error('terminateVmBySession: Failed to kill PID %u (VM %s)' % (uPid, oSession.sName));
 
         # The final steps.
         if fClose is True:
+            reporter.log('terminateVmBySession: closing session "%s"...' % (oSession.sName,));
             oSession.close();
             self.waitOnDirectSessionClose(oSession.oVM, 10000);
             try:
